+++ src/main/java/com/recipe/app/src/common/infra/NoticeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface NoticeRepository extends JpaRepository<Notice, Integer> {
+public interface NoticeRepository extends JpaRepository<Notice, Long> {
 
     Optional<Notice> findFirstByActiveYn(String activeYn);
 }
